+++ utils/editor/src/models/data/deploy.ts
@@ -1,0 +1,9 @@
+export interface DeployData {
+  status: 'scheduled' | 'failure' | 'successful';
+  updated_at: Date | number | BigInt;
+}
+
+export interface Deploy {
+  github?: DeployData;
+  api?: DeployData;
+}